[feat] 평가 결과 전송 구현

+++ app/src/main/java/com/ssu/bilda/data/remote/request/EvaluationRequest.kt
@@ -3,8 +3,7 @@ package com.ssu.bilda.data.remote.request
 import Scores
 
 data class EvaluationRequest(
-    val evaluatedUserId: Long,
-    val teamId: Long,
+    val evaluatedUserId: Int,
+    val teamId: Int,
     val scores: Scores
 )
-
